@@ -1,7 +1,7 @@
 {
   "name": "base58-universal",
   "version": "1.0.1-0",
-  "description": "Encode/decode using \"The Base58 Encoding Scheme\".",
+  "description": "Encode/decode using the Base58 Bitcoin Encoding Scheme.",
   "homepage": "https://github.com/digitalbazaar/base58-universal",
   "author": {
     "name": "Digital Bazaar, Inc.",
@@ -17,26 +17,33 @@
     "email": "support@digitalbazaar.com"
   },
   "license": "BSD-3-Clause",
-  "main": "lib/index.js",
-  "module": "lib/main.js",
   "files": [
-    "lib/"
+    "dist",
+    "lib",
+    "README.md",
+    "LICENSE"
   ],
-  "dependencies": {
-    "esm": "^3.2.25"
+  "main": "dist/index.js",
+  "browser": "dist/index.js",
+  "module": "dist/esm/index.js",
+  "exports": {
+    ".": {
+      "require": "./dist/index.js",
+      "import": "./dist/esm/index.js"
+    }
   },
+  "dependencies": {},
   "devDependencies": {
     "@babel/core": "^7.14.0",
-    "@babel/plugin-transform-modules-commonjs": "^7.8.3",
     "@babel/plugin-transform-runtime": "^7.8.3",
     "@babel/preset-env": "^7.14.1",
     "@babel/runtime": "^7.8.3",
     "babel-loader": "^8.0.6",
-    "chai": "^4.2.0",
-    "chai-bytes": "^0.1.2",
-    "cross-env": "^6.0.3",
-    "eslint": "^6.8.0",
-    "eslint-config-digitalbazaar": "^2.1.0",
+    "chai": "^4.3.3",
+    "cross-env": "^7.0.3",
+    "eslint": "^7.21.0",
+    "eslint-config-digitalbazaar": "^2.6.1",
+    "esm": "^3.2.25",
     "karma": "^4.4.1",
     "karma-babel-preprocessor": "^8.0.1",
     "karma-chrome-launcher": "^3.1.0",
@@ -51,6 +58,8 @@
     "mocha": "^7.0.0",
     "mocha-lcov-reporter": "^1.3.0",
     "nyc": "^15.0.0",
+    "rimraf": "^3.0.2",
+    "rollup": "^2.47.0",
     "webpack": "^4.41.5"
   },
   "nyc": {
@@ -66,12 +75,16 @@
     "node": ">=12"
   },
   "scripts": {
-    "test": "npm run test-node",
-    "test-node": "cross-env NODE_ENV=test mocha --preserve-symlinks -t 30000 -A -R ${REPORTER:-spec} test/*.spec.js",
-    "test-karma": "karma start",
-    "benchmark": "node benchmark/benchmark.js",
+    "rollup": "rollup -c rollup.config.js",
+    "build": "npm run clear && npm run rollup && ./build-dist.sh",
+    "clear": "rimraf dist/*",
     "coverage": "cross-env NODE_ENV=test nyc --reporter=lcov --reporter=text-summary npm run test-node",
     "coverage-report": "nyc report",
+    "prepare": "npm run build",
+    "test": "npm run test-node",
+    "rebuild": "npm run clear && npm run build",
+    "test-node": "cross-env NODE_ENV=test mocha --preserve-symlinks -r esm -t 30000 -A -R ${REPORTER:-spec} test/*.spec.js",
+    "test-karma": "karma start",
     "lint": "eslint ."
   },
   "keywords": [
