@@ -30,6 +30,9 @@
  */
 'use strict';
 
+// base58 characters (Bitcoin alphabet)
+const BASE58_BTC = '123456789ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz';
+
 // baseN alphabet indexes
 const _reverseAlphabets = {};
 
@@ -42,7 +45,7 @@ const _reverseAlphabets = {};
  *
  * @return {string} the baseN-encoded output string.
  */
-export function encode(input, alphabet, maxline) {
+export function encode(input, maxline, alphabet = BASE58_BTC) {
   if(!(input instanceof Uint8Array)) {
     throw new TypeError('"input" must be a Uint8Array.');
   }
@@ -101,7 +104,7 @@ export function encode(input, alphabet, maxline) {
  *
  * @return {Uint8Array} the decoded bytes in a Uint8Array.
  */
-export function decode(input, alphabet) {
+export function decode(input, alphabet = BASE58_BTC) {
   if(typeof input !== 'string') {
     throw new TypeError('"input" must be a string.');
   }
